user/gitlay: remove bundled git
gitaly bundles two versions of git (v2.35.1 and v2.37.1).
If git.use_bundled_binaries is enabled and there are no *both* of these
versions available, it fails to start. This is stupid.

Moreover, use_bundled_binaries is not even used by default (yet) and it doesn't
seem to be any problem with using a newer git version. It was a bad idea to
package the bundled git in the first place, it was probably too late and too
much wine...

doc/git-execution-environments.md:
> While we backport patches that have been accepted in the upstream Git project,
> we do not plan to apply patches that will not eventually end up in Git itself.